requirements.txt
attrs==23.1.0
Automat==22.10.0
azure-ai-textanalytics==5.2.1
azure-common==1.1.28
azure-core==1.26.4
azure-data-tables==12.4.2
azure-functions==1.14.0
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
filelock==3.12.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.6.2
urllib3==2.0.2
w3lib==2.1.1
yarl==1.9.2
zope.interface==6.0
